+++ core/parser/__init__.py
@@ -1,8 +1,13 @@
 from core.parser.diff import ChangedFile, parse_changed_files
+from core.parser.python_ast import PythonAstParser, PythonAstSummary, PythonImport, PythonSymbol
 from core.parser.repository import RepositoryFile, RepositorySnapshot, RepositoryStructureParser
 
 __all__ = [
     "ChangedFile",
+    "PythonAstParser",
+    "PythonAstSummary",
+    "PythonImport",
+    "PythonSymbol",
     "RepositoryFile",
     "RepositorySnapshot",
     "RepositoryStructureParser",